refactor: login, dashboard, register pages
