+++ nextjs-client/src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Header from "./layout/header";
+import Header from "../components/Header";
 
 const inter = Inter({ subsets: ["latin"] });
