+++ nextjs-client/src/app/loading.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import LoadingSpinner from "./layout/loadingSpinner";
+import LoadingSpinner from "@/src/components/Loaders/LoadingSpinner";
 
 export default function Loading() {
   return <LoadingSpinner />;
