+++ nextjs-client/src/components/Header/index.tsx
@@ -1,32 +1,27 @@
 "use client";
 import headerLock from "@/public/assets/lockHeader.png";
 import profilePhoto from "@/public/assets/user-profile.png";
-import { Cookies } from "react-cookie";
 import Link from "next/link";
-import { redirect } from "next/navigation";
+import { redirect, useRouter } from "next/navigation";
 import Image from "next/image";
+import useAuthStore from "@/src/store/authStore";
+import useProfileStore from "@/src/store/profileStore";
+import { deleteCookie } from "cookies-next";
 
 export default function Header() {
-  const cookies = new Cookies();
-  const cookieValue = cookies.get("myCookie");
-  const userName = cookieValue?.userName;
+  const router = useRouter();
+  const token = useAuthStore((state) => state.authToken);
+  const profile = useProfileStore((state) => state.profileDetails);
 
-  const logout = () => {
-    const fetchData = async () => {
-      // get the data from the api
-      await fetch(`${process.env.REACT_APP_HOST_URL}/auth/logout`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        credentials: "include",
-      });
-    };
-    fetchData();
-    cookies.remove("myCookie", { path: "/" });
+  const setAuthToken = useAuthStore((state) => state.setAuthToken);
+  const setProfile = useProfileStore((state) => state.setProfileDetails);
 
-    window.localStorage.clear();
-    redirect("/");
+  const logout = async () => {
+    deleteCookie("token");
+    router.replace("/auth/login");
+
+    setAuthToken("");
+    setProfile({});
   };
 
   return (
@@ -34,7 +29,7 @@ export default function Header() {
       data-testid="header"
       className=" bg-cyan-500 w-full text-white  px-4 flex flex-row justify-between items-center h-16"
     >
-      {!userName ? (
+      {!token ? (
         <div className="flex flex-row w-full justify-between items-center">
           <div className="flex flew-row items-end">
             <Image className="w-6 h-8" src={headerLock} alt="" />
@@ -58,7 +53,7 @@ export default function Header() {
             <div className="flex flex-row">
               <div className="">
                 <Image className="" src={profilePhoto} alt="" />
-                <div className=""> {userName} </div>
+                <div className=""> {profile.name} </div>
               </div>
               <button className="border p-2 rounded" onClick={logout}>
                 Logout
