+++ nextjs-client/src/components/Loaders/LoadingSpinner/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../styles/LoadingSpinner.css"
+import "@/src/styles/LoadingSpinner.css";
 
 export default function LoadingSpinner() {
   return (
