+++ nextjs-client/src/components/Passwords/deleteConfirmation.tsx
